ai_service/requirements.txt
fastapi
uvicorn
pillow
numpy<2
tflite-runtime
